requirements_clean.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
dj-database-url==2.2.0
whitenoise==6.8.2
requests==2.32.5
celery==5.5.3
redis==5.2.1
channels==4.3.1
channels-redis==4.3.0
arrow==1.4.0
asgiref==3.10.0
billiard==4.2.2
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
idna==3.11
Jinja2==3.1.6
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
packaging==25.0
prompt_toolkit==3.0.52
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
rich==14.2.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
inflection>=0.5.1
uritemplate>=4.1.1
